swctl: switch back to 115200 bauds before boot.
swctl changes the UART speed to upload the test binary faster but until
now the kernel also had to be configured with this high speed rate to
get the console.

Now the baud rate is set back to 115200 after the upload of the kernel
image. It allows to boot a kernel with a usual 115200 baudrate.